// src/pages/Enroll/ChoiceTeam/data.d.ts
export interface EnrollTeamData {
  status: string;
  id: number;
  matchdata: number;
  unitdata: number;
  name: string;
  leader: string;
  leaderphonenumber: string;
  email: string;
  coachone: string;
  coachonephonenumber: string;
  coachtwo: string;
  coachtwophonenumber: string;
  dutybook: string;
  url_dutybook: string;
}